requirements.txt
certifi==2020.11.8
colorama==0.4.4
crayons==0.3.1
cycler==0.10.0
image-slicer==0.3.0
kiwisolver==1.3.1
matplotlib==3.3.2
numpy==1.19.2
piexif==1.1.3
Pillow==6.2.2
pycrypto==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scipy==1.5.4
six==1.15.0
stegano==0.9.8
